batch run of pure ddpg with different seeds

+++ msc_project/algorithms/ddpg/buffer.py
@@ -5,10 +5,10 @@ class ReplayBuffer:
     def __init__(self, max_size, input_shape, n_actions):
         self.mem_size = max_size
         self.mem_cntr = 0
-        self.state_memory = np.zeros((self.mem_size, input_shape))
-        self.new_state_memory = np.zeros((self.mem_size, input_shape))
-        self.action_memory = np.zeros((self.mem_size, n_actions))
-        self.reward_memory = np.zeros(self.mem_size)
+        self.state_memory = np.zeros((self.mem_size, input_shape), dtype=np.float32)
+        self.new_state_memory = np.zeros((self.mem_size, input_shape), dtype=np.float32)
+        self.action_memory = np.zeros((self.mem_size, n_actions), dtype=np.float32)
+        self.reward_memory = np.zeros(self.mem_size, dtype=np.float32)
         self.terminal_memory = np.zeros(self.mem_size, dtype=np.bool)
 
     def store_transition(self, state, action, reward, state_, done):
